Convert function generate_and_print_maps to generate_maps
